перенес логику создания фотографий в отдельный модуль
