+++ js/util.js
@@ -21,5 +21,7 @@ const createRandomId = (min, max) => {
   };
 };
 
-export { getRandomInteger };
-export { createRandomId };
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export { getRandomInteger, createRandomId, getRandomArrayElement };
+
